Allow spacers to have Tasks

+++ Marvin/src/main/java/kutch/biff/marvin/utility/FrameworkNode.java
@@ -733,7 +733,8 @@ public class FrameworkNode
             }
 
             retString = strData.substring(0, OutterIndex);
-            retString += Integer.toString((int) NewVal);
+            //retString += Integer.toString((int) NewVal);
+            retString += Double.toString(NewVal);
             retString += strData.substring(CloseParenIndex + 1);
         }
         else
